requirements.txt
# Core LangChain and LangGraph dependencies
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# LLM providers
openai>=1.0.0

# Search and web tools
tavily-python>=0.5.0
httpx>=0.27.0
markdownify>=0.13.0

# Data handling
pydantic>=2.0.0

# Utility packages
python-dotenv>=1.0.0
rich>=13.0.0

# Visualization
matplotlib>=3.7.0
numpy>=1.24.0

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: For enhanced functionality
typing-extensions>=4.0.0